SEQUOIADBMAINSTREAM-4368 清理队列每次都是向后增加新的桶名称，即使有重复的名称也会加，测试用例频繁尝试修改，导致队列中堆积了大量的重复桶名称，多个用例并发时，有可能最后一个桶要等很长时间才能等到处理清理动作。 修改为存储不重复的桶名称，一个桶只有一条记录在队列中
git-svn-id: http://192.168.20.11/sequoiadb/trunk@40472 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/sequoias3/src/main/java/com/sequoias3/taskmanager/DelimiterClean.java
@@ -58,8 +58,8 @@ public class DelimiterClean {
                 region = regionDao.queryRegion(bucket.getRegion());
             }
             String metaCsName = regionDao.getMetaCurCSName(region);
-            if (bucket.getDelimiter1Status().equals(DelimiterStatus.DELETING.getName())
-                    || bucket.getDelimiter2Status().equals(DelimiterStatus.DELETING.getName())){
+            if ((bucket.getDelimiter1Status() != null && bucket.getDelimiter1Status().equals(DelimiterStatus.DELETING.getName()))
+                    || (bucket.getDelimiter2Status() != null && bucket.getDelimiter2Status().equals(DelimiterStatus.DELETING.getName()))){
                 ConnectionDao connectionDao = daoMgr.getConnectionDao();
                 transaction.begin(connectionDao);
                 try {
@@ -104,18 +104,18 @@ public class DelimiterClean {
                 }
             }
         }catch (Exception e){
-            logger.error("clean deleting bucket failed. bucketName:"+bucketName);
+            logger.error("clean deleting bucket failed. bucketName:"+bucketName, e);
         }
     }
 
     private void cleanDelimiter(ConnectionDao connectionDao, String metaCsName, Bucket bucket)
             throws S3ServerException {
         Integer delimiterFlag = null;
-        if (bucket.getDelimiter1Status().equals(DelimiterStatus.DELETING.getName())){
+        if (bucket.getDelimiter1Status() != null && bucket.getDelimiter1Status().equals(DelimiterStatus.DELETING.getName())){
             dirDao.delete(connectionDao, metaCsName, bucket.getBucketId(), bucket.getDelimiter1(), null);
             delimiterFlag = 1;
         }
-        if (bucket.getDelimiter2Status().equals(DelimiterStatus.DELETING.getName())){
+        if (bucket.getDelimiter2Status() != null && bucket.getDelimiter2Status().equals(DelimiterStatus.DELETING.getName())){
             dirDao.delete(connectionDao, metaCsName, bucket.getBucketId(), bucket.getDelimiter2(), null);
             delimiterFlag = 2;
         }
